combine classes
