+++ platform-sdk/consensus-model/build.gradle.kts
@@ -9,7 +9,6 @@ plugins {
 description = "Consensus Model"
 
 testModuleInfo {
-    requires("com.swirlds.platform.core")
     requires("com.swirlds.platform.core.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.model.test.fixtures")
