+++ platform-sdk/consensus-model/src/test/java/org/hiero/consensus/model/event/EventSerializationUtils.java
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+package org.hiero.consensus.model.event;
+
+/**
+ * A utility class for serializing and deserializing events using legacy serialization. This will be removed as soon as
+ * event serialization is completely migrated to protobuf.
+ */
+public final class EventSerializationUtils {
+    private EventSerializationUtils() {
+        // Utility class
+    }
+}
