+++ platform-sdk/swirlds-platform-core/src/main/java/module-info.java
@@ -59,7 +59,6 @@ module com.swirlds.platform.core {
     exports com.swirlds.platform.event.branching to
             org.hiero.consensus.reconnect.impl;
     exports com.swirlds.platform.reconnect;
-    exports com.swirlds.platform.event;
     exports com.swirlds.platform.state.nexus to
             org.hiero.consensus.reconnect.impl;
     exports com.swirlds.platform.wiring;
